# Control_Project_JSON/convenience.store.api-main/src/main/java/com/conveniencestore/conveniencestore/domain/users/EditUserDTO.java
package com.conveniencestore.conveniencestore.domain.users;

import jakarta.validation.constraints.Email;

public record EditUserDTO(
        String username,
        @Email
        String email
) {
}